refactor: Moved code for calculating in range IDs to `data_utils`
